# docs/swagger.yaml
...
swagger: "2.0"
info:
  description: "API of Metadata Service for LAAWS"
  version: "1.0.0"
  title: "LAAWS Metadata Service API"
  termsOfService: "http://www.lockss.org/terms/"
  contact:
    name: "LOCKSS Support"
    url: "http://www.lockss.org/"
    email: "support@lockss.org"
  license:
    name: "Modified BSD License"
    url: "http://www.lockss.org/support/open-source-license/"
host: "laaws.lockss.org:443"
basePath: "/"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- basicAuth: []
paths:
  /metadata/aus:
    post:
      tags:
      - "metadata"
      summary: "Store the metadata for an AU item"
      description: "Store the metadata for an item belonging to an AU"
      operationId: "postMetadataAusItem"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "item"
        in: "body"
        description: "The metadata of the AU item to be stored"
        required: true
        schema:
          $ref: "#/definitions/itemMetadata"
      responses:
        200:
          description: "The key under which the metadata of the AU item has\
            \ been stored"
          schema:
            type: "integer"
            format: "int64"
        401:
          description: "Unauthorized request"
        403:
          description: "Forbidden request"
        500:
          description: "Internal server error"
        503:
          description: "Some or all of the system is not available"
  /metadata/aus/{auid}:
    get:
      tags:
      - "metadata"
      summary: "Get the metadata stored for an AU"
      description: "Get the full metadata stored for an AU given the AU\
        \ identifier or a pageful of the metadata defined by the page index and\
        \ size"
      operationId: "getMetadataAusAuid"
      produces:
      - "application/json"
      parameters:
      - name: "auid"
        in: "path"
        description: "The identifier of the AU for which the metadata is\
          \ requested"
        required: true
        type: "string"
      - name: "page"
        in: "query"
        required: false
        type: "integer"
        description: "The identifier of the page of metadata to be returned"
        default: 1
      - name: "limit"
        in: "query"
        required: false
        type: "integer"
        description: "The number of items per page"
        default: 50
      responses:
        200:
          description: "The metadata of the specified AU"
          schema:
            $ref: "#/definitions/auMetadataPageInfo"
        401:
          description: "Unauthorized request"
        404:
          description: "AU not found"
        500:
          description: "Internal server error"
        503:
          description: "Some or all of the system is not available"
    delete:
      tags:
      - "metadata"
      summary: "Delete the metadata stored for an AU"
      description: "Delete the metadata stored for an AU given the AU identifier"
      operationId: "deleteMetadataAusAuid"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "auid"
        in: "path"
        description: "The identifier of the AU for which the metadata is to be\
          \ deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "The number of deleted AU metadata items"
          schema:
            type: "integer"
        401:
          description: "Unauthorized request"
        403:
          description: "Forbidden request"
        404:
          description: "AU not found"
        500:
          description: "Internal server error"
        503:
          description: "Some or all of the system is not available"
...
securityDefinitions:
  basicAuth:
    description: "HTTP Basic Authentication. Works over `HTTP` and `HTTPS`"
    type: "basic"
definitions:
  auMetadataPageInfo:
    type: "object"
    required:
    - "items"
    - "pageInfo"
    properties:
      items:
        type: "array"
        description: "The metadata for the AU items in the page"
        items:
          $ref: "#/definitions/itemMetadata"
      pageInfo:
        $ref: "#/definitions/pageInfo"
    description: "The display page of AU metadata"
  itemMetadata:
    type: "object"
    properties:
      scalarMap:
        type: "object"
        description: "The map of scalar metadata elements for this item"
        additionalProperties:
          type: "string"
          description: "The scalar value of the metadata for the key"
      listMap:
        type: "object"
        description: "The map of listed metadata elements for this item"
        additionalProperties:
          type: "array"
          description: "The list value of the metadata for the key"
          items:
            type: "string"
      mapMap:
        type: "object"
        description: "The map of mapped metadata elements for this item"
        additionalProperties:
          type: "object"
          description: "The map value of the metadata for the key"
          additionalProperties:
            type: "string"
    description: "The metadata generated for a single item"
  pageInfo:
    type: "object"
    required:
    - "curLink"
    - "currentPage"
    - "resultsPerPage"
    - "totalCount"
    properties:
      totalCount:
        type: "integer"
        format: "int32"
        description: "The total number of elements to be paginated"
      resultsPerPage:
        type: "integer"
        format: "int32"
        description: "The number of results per page"
      currentPage:
        type: "integer"
        format: "int32"
        description: "The current page number"
      curLink:
        type: "string"
        description: "The link to the current page"
      nextLink:
        type: "string"
        description: "The link to the next page"
    description: "The information related to pagination of content"
...